+++ src/main/java/eu/sedov/servlet/DeleteUserServlet.java
@@ -2,7 +2,7 @@ package eu.sedov.servlet;
 
 import eu.sedov.model.User;
 import eu.sedov.repository.mapper.impl.UserEnumMap;
-import eu.sedov.service.UserService;
+import eu.sedov.service.LibraryUserService;
 import eu.sedov.servlet.dto.OutUserDTO;
 import eu.sedov.servlet.mapper.UserMapperDto;
 
@@ -15,22 +15,23 @@ import java.io.IOException;
 
 @WebServlet("/delete")
 public class DeleteUserServlet extends HttpServlet {
-    private UserService service;
+    private LibraryUserService service;
     private UserMapperDto mapper;
-    private UserEnumMap map;
+    private UserEnumMap userEnum;
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        UserEnumMap map = new UserEnumMap();
-        int id = Integer.parseInt(req.getParameter(map.getMap().get(UserEnumMap.UserResultSetParams.ID)));
+        var stringId = req.getParameter(userEnum.getMap().get(UserEnumMap.UserResultSetParams.ID));
+        if (stringId != null){
+            Integer id =Integer.parseInt(stringId);
+            User user = service.getById(id);
+            int deletedUsersCount = service.delete(id);
 
-        User user = service.getById(id);
-        int deletedUsersCount = service.delete(id);
-
-        if (deletedUsersCount != 0){
-            OutUserDTO deletedUserDto = mapper.map(user);
-            req.setAttribute("deletedUser", deletedUserDto);
+            if (deletedUsersCount != 0){
+                OutUserDTO deletedUserDto = mapper.map(user);
+                req.setAttribute("deletedUser", deletedUserDto);
+            }
         }
 
         req.getRequestDispatcher("/delete").forward(req, resp);
